requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.6
distlib==0.3.8
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Faker==25.9.1
filelock==3.15.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
platformdirs==4.2.2
pylint==2.5.3
python-dateutil==2.9.0.post0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.5.5
tzdata==2024.1
validate-docbr==1.10.0
virtualenv==20.26.2
wrapt==1.12.1
